메인화면 상속 기능 구현
