+++ app/src/main/AndroidManifest.xml
@@ -19,6 +19,7 @@
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+
         </activity>
         <activity
             android:name=".sidemenu_baseactivity"
